requirements.txt
black==21.9b0
boto3==1.18.63
botocore==1.21.63
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
ecdsa==0.17.0
envs==1.3
future==0.18.2
idna==3.3
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycryptodome==3.3.1
python-dateutil==2.8.2
python-dotenv==0.19.1
python-jose-cryptodome==1.3.2
regex==2021.10.8
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
warrant==0.6.1
